requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
psycopg[binary]==3.2.10
python-dotenv==1.1.1
jinja2==3.1.6
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-telegram-bot==22.3
